修改统计数据

+++ src/main/java/com/chebianjie/datacleaning/domain/UtChargeLog.java
@@ -45,7 +45,7 @@ public class UtChargeLog implements Serializable {
      */
     @ApiModelProperty(value = "获取人uid(旧系统的ims_jy_cbj2_user.uid，新系统的ut_consumer.uid)，为了兼容旧系统而保留的uid")
     @Column(name = "consumer_id")
-    private Integer consumerId;
+    private Long consumerId;
 
 
     /** 消费者账号 */
